+++ src/module/core/ics05_port.rs
@@ -1,4 +1,4 @@
-use crate::{utils::LOG_TARGET, *};
+use crate::*;
 use log::trace;
 
 use crate::context::Context;
